Now removes the synopsis before parsing MAL XML
MAL doesn't give nice XML responses, and the synopsis tags often had unescaped
HTML entities that broke the XML parser. My previous solution was to try to parse it once, then
escape each HTML entities if it failed (Since it is a lot of search and replace calls).
Now I remove the synopsis in a since replaceAll call.